create ValheimBackupShared project for business logic, and ValheimBackupService for the backup service. Move application settings to the shared project instead to avoid a circular reference between the ui project and shared project
