+++ .gitignore
@@ -1,3 +1,3 @@
 packages/
-ValheimBackup/bin/
-ValheimBackup/obj/
+*/bin/
+*/obj/
