+++ ValheimBackupService/ValheimBackupService.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Diagnostics;
+using System.Linq;
+using System.ServiceProcess;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ValheimBackupService
+{
+    public partial class ValheimBackupService : ServiceBase
+    {
+        public ValheimBackupService()
+        {
+            InitializeComponent();
+        }
+
+        protected override void OnStart(string[] args)
+        {
+        }
+
+        protected override void OnStop()
+        {
+        }
+    }
+}
