+++ ValheimBackupShared/BO/FtpConnectionInfo.cs
