+++ ValheimBackupShared/FTP/FtpFileInfo.cs
