+++ ValheimBackupShared/Properties/Settings.settings
@@ -1,0 +1,18 @@
+﻿<?xml version='1.0' encoding='utf-8'?>
+<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="ValheimBackup.Properties" GeneratedClassName="Settings">
+  <Profiles />
+  <Settings>
+    <Setting Name="ServersFileName" Type="System.String" Scope="Application">
+      <Value Profile="(Default)">servers.json</Value>
+    </Setting>
+    <Setting Name="BackupsFileName" Type="System.String" Scope="Application">
+      <Value Profile="(Default)">backups.json</Value>
+    </Setting>
+    <Setting Name="AppDataDirectory" Type="System.String" Scope="User">
+      <Value Profile="(Default)" />
+    </Setting>
+    <Setting Name="DefaultBackupDirectory" Type="System.String" Scope="User">
+      <Value Profile="(Default)" />
+    </Setting>
+  </Settings>
+</SettingsFile>
